Small modifications for d2
